# release-19.2: bench: use round-robin for MultinodeCockroach