--- a/pkg/bench/pgbench_test.go
+++ b/pkg/bench/pgbench_test.go
@@ -12,7 +12,6 @@ package bench
 
 import (
 	"context"
-	gosql "database/sql"
 	"fmt"
 	"math/rand"
 	"net"
@@ -33,14 +32,14 @@ import (
 // in its TPC-B(ish) mode.
 func BenchmarkPgbenchQuery(b *testing.B) {
 	defer log.Scope(b).Close(b)
-	ForEachDB(b, func(b *testing.B, db *gosql.DB) {
-		if err := SetupBenchDB(db, 20000, true /*quiet*/); err != nil {
+	ForEachDB(b, func(b *testing.B, db *sqlutils.SQLRunner) {
+		if err := SetupBenchDB(db.DB, 20000, true /*quiet*/); err != nil {
 			b.Fatal(err)
 		}
 		src := rand.New(rand.NewSource(5432))
 		b.ResetTimer()
 		for i := 0; i < b.N; i++ {
-			if err := RunOne(db, src, 20000); err != nil {
+			if err := RunOne(db.DB, src, 20000); err != nil {
 				b.Fatal(err)
 			}
 		}
@@ -52,8 +51,8 @@ func BenchmarkPgbenchQuery(b *testing.B) {
 // in its TPC-B(ish) mode.
 func BenchmarkPgbenchQueryParallel(b *testing.B) {
 	defer log.Scope(b).Close(b)
-	ForEachDB(b, func(b *testing.B, db *gosql.DB) {
-		if err := SetupBenchDB(db, 20000, true /*quiet*/); err != nil {
+	ForEachDB(b, func(b *testing.B, db *sqlutils.SQLRunner) {
+		if err := SetupBenchDB(db.DB, 20000, true /*quiet*/); err != nil {
 			b.Fatal(err)
 		}
 
@@ -71,7 +70,7 @@ func BenchmarkPgbenchQueryParallel(b *testing.B) {
 			for pb.Next() {
 				r.Reset()
 				for r.Next() {
-					err = RunOne(db, src, 20000)
+					err = RunOne(db.DB, src, 20000)
 					if err == nil {
 						break
 					}