--- a/pkg/bench/setup.go
+++ b/pkg/bench/setup.go
@@ -12,10 +12,13 @@ package bench
 
 import (
 	"bytes"
+	"context"
 	gosql "database/sql"
 	"fmt"
 	"net/url"
 	"os/exec"
+
+	"github.com/cockroachdb/cockroach/pkg/testutils/sqlutils"
 )
 
 const schema = `
@@ -105,18 +108,19 @@ func SetupExec(pgURL url.URL, name string, accounts, transactions int) (*exec.Cm
 // not support. The queries this script runs are based on a dump of a db created
 // by `pgbench -i`, but sticking to the compatible subset that both cockroach and
 // postgres support.
-func SetupBenchDB(db *gosql.DB, accounts int, quiet bool) error {
-	if _, err := db.Exec(schema); err != nil {
+func SetupBenchDB(db sqlutils.DBHandle, accounts int, quiet bool) error {
+	ctx := context.TODO()
+	if _, err := db.ExecContext(ctx, schema); err != nil {
 		return err
 	}
-	return populateDB(db, accounts, quiet)
+	return populateDB(ctx, db, accounts, quiet)
 }
 
 const tellers = 10
 
-func populateDB(db *gosql.DB, accounts int, quiet bool) error {
+func populateDB(ctx context.Context, db sqlutils.DBHandle, accounts int, quiet bool) error {
 	branches := `INSERT INTO pgbench_branches (bid, bbalance, filler) VALUES (1, 7354, NULL)`
-	if r, err := db.Exec(branches); err != nil {
+	if r, err := db.ExecContext(ctx, branches); err != nil {
 		return err
 	} else if x, err := r.RowsAffected(); err != nil {
 		return err
@@ -136,7 +140,7 @@ func populateDB(db *gosql.DB, accounts int, quiet bool) error {
 	(9, 1, 0, NULL),
 	(10, 1, 3345, NULL)
 	`
-	if r, err := db.Exec(tellers); err != nil {
+	if r, err := db.ExecContext(ctx, tellers); err != nil {
 		return err
 	} else if x, err := r.RowsAffected(); err != nil {
 		return err
@@ -163,7 +167,7 @@ func populateDB(db *gosql.DB, accounts int, quiet bool) error {
 			fmt.Fprintf(&placeholders, "(%d, 1, 0, '                                                                                    ')", done+i)
 		}
 		stmt := fmt.Sprintf(`INSERT INTO pgbench_accounts VALUES %s`, placeholders.String())
-		if r, err := db.Exec(stmt); err != nil {
+		if r, err := db.ExecContext(ctx, stmt); err != nil {
 			return err
 		} else if x, err := r.RowsAffected(); err != nil {
 			return err
@@ -186,7 +190,7 @@ INSERT INTO pgbench_history VALUES
 (5, 1, 68648, 1880, CURRENT_TIMESTAMP, NULL),
 (10, 1, 46989, 1080, CURRENT_TIMESTAMP, NULL);`
 
-	if r, err := db.Exec(history); err != nil {
+	if r, err := db.ExecContext(ctx, history); err != nil {
 		return err
 	} else if x, err := r.RowsAffected(); err != nil {
 		return err