--- a/pkg/sql/tests/bank_test.go
+++ b/pkg/sql/tests/bank_test.go
@@ -12,12 +12,12 @@ package tests_test
 
 import (
 	"bytes"
-	gosql "database/sql"
 	"fmt"
 	"math/rand"
 	"testing"
 
 	"github.com/cockroachdb/cockroach/pkg/bench"
+	"github.com/cockroachdb/cockroach/pkg/testutils/sqlutils"
 )
 
 // maxTransfer is the maximum amount to transfer in one transaction.
@@ -26,7 +26,7 @@ const maxTransfer = 999
 // runBenchmarkBank mirrors the SQL performed by examples/sql_bank, but
 // structured as a benchmark for easier usage of the Go performance analysis
 // tools like pprof, memprof and trace.
-func runBenchmarkBank(b *testing.B, db *gosql.DB, numAccounts int) {
+func runBenchmarkBank(b *testing.B, db *sqlutils.SQLRunner, numAccounts int) {
 	{
 		// Initialize the "bank" table.
 		schema := `
@@ -34,12 +34,8 @@ CREATE TABLE IF NOT EXISTS bench.bank (
   id INT PRIMARY KEY,
   balance INT NOT NULL
 )`
-		if _, err := db.Exec(schema); err != nil {
-			b.Fatal(err)
-		}
-		if _, err := db.Exec("TRUNCATE TABLE bench.bank"); err != nil {
-			b.Fatal(err)
-		}
+		db.Exec(b, schema)
+		db.Exec(b, "TRUNCATE TABLE bench.bank")
 
 		var placeholders bytes.Buffer
 		var values []interface{}
@@ -51,9 +47,7 @@ CREATE TABLE IF NOT EXISTS bench.bank (
 			values = append(values, i)
 		}
 		stmt := `INSERT INTO bench.bank (id, balance) VALUES ` + placeholders.String()
-		if _, err := db.Exec(stmt, values...); err != nil {
-			b.Fatal(err)
-		}
+		db.Exec(b, stmt, values...)
 	}
 
 	b.ResetTimer()
@@ -71,16 +65,14 @@ CREATE TABLE IF NOT EXISTS bench.bank (
 UPDATE bench.bank
   SET balance = CASE id WHEN $1 THEN balance-$3 WHEN $2 THEN balance+$3 END
   WHERE id IN ($1, $2) AND (SELECT balance >= $3 FROM bench.bank WHERE id = $1)`
-			if _, err := db.Exec(update, from, to, amount); err != nil {
-				b.Fatal(err)
-			}
+			db.Exec(b, update, from, to, amount)
 		}
 	})
 	b.StopTimer()
 }
 
 func BenchmarkBank(b *testing.B) {
-	bench.ForEachDB(b, func(b *testing.B, db *gosql.DB) {
+	bench.ForEachDB(b, func(b *testing.B, db *sqlutils.SQLRunner) {
 		for _, numAccounts := range []int{2, 4, 8, 32, 64} {
 			b.Run(fmt.Sprintf("numAccounts=%d", numAccounts), func(b *testing.B) {
 				runBenchmarkBank(b, db, numAccounts)