Backport 1/1 commits from #41590.

/cc @cockroachdb/release

---

When these benchmarks run against MultinodeCockroach, they only talk
to the first node in a three-node cluster. This is unrealistic and
leads to silly behaviors where sub-benchmarks (which run against the
same cluster) change performance when a relevant lease moves.

This change makes it so that we issue queries against all three nodes,
in round-robin fashion. This required some refactoring; we add the
round-robin capability to `SQLRunner` and use that everywhere.

Note that the results of the Multinode benchmarks going forward will
not be directly comparable with previous results.

```
name                                                                                 old time/op  new time/op   delta
Select1/Cockroach-24                                                                  167µs ± 4%    168µs ± 3%      ~     (p=0.549 n=10+9)
Select1/MultinodeCockroach-24                                                         171µs ± 2%    176µs ± 1%    +2.93%  (p=0.000 n=8+8)
Select2/Cockroach-24                                                                  866µs ± 1%    856µs ± 1%    -1.08%  (p=0.002 n=10+10)
Select2/MultinodeCockroach-24                                                         875µs ± 0%   1547µs ± 1%   +76.88%  (p=0.000 n=10+10)
Select3/Cockroach-24                                                                 1.15ms ±12%   1.10ms ± 6%      ~     (p=0.113 n=10+9)
Select3/MultinodeCockroach-24                                                        1.10ms ± 1%   1.82ms ± 1%   +65.22%  (p=0.000 n=10+10)
Count/Cockroach-24                                                                   25.7ms ± 1%   25.4ms ± 1%    -1.22%  (p=0.000 n=9+9)
Count/MultinodeCockroach-24                                                          25.5ms ± 1%   26.2ms ± 1%    +2.75%  (p=0.000 n=9+8)
CountTwoCF/Cockroach-24                                                              45.4ms ± 1%   44.8ms ± 1%    -1.37%  (p=0.000 n=10+10)
CountTwoCF/MultinodeCockroach-24                                                     45.5ms ± 1%   45.8ms ± 1%      ~     (p=0.156 n=10+9)
Sort/Cockroach-24                                                                     156ms ± 1%    157ms ± 1%      ~     (p=0.739 n=10+10)
Sort/MultinodeCockroach-24                                                            156ms ± 1%    206ms ± 1%   +32.16%  (p=0.000 n=10+9)
SQL/Cockroach/Delete/count=1-24                                                       556µs ± 3%    546µs ± 2%    -1.87%  (p=0.004 n=9+9)
SQL/Cockroach/Delete/count=10-24                                                      660µs ± 6%    675µs ±11%      ~     (p=0.631 n=10+10)
SQL/Cockroach/Delete/count=100-24                                                    1.41ms ± 5%   1.39ms ± 4%      ~     (p=0.161 n=9+9)
SQL/Cockroach/Delete/count=1000-24                                                   19.4ms ±15%   19.4ms ±14%      ~     (p=0.853 n=10+10)
SQL/Cockroach/Insert/count=1-24                                                       600µs ± 2%    592µs ± 3%      ~     (p=0.063 n=10+10)
SQL/Cockroach/Insert/count=10-24                                                      744µs ± 1%    744µs ± 3%      ~     (p=0.661 n=9+10)
SQL/Cockroach/Insert/count=100-24                                                    1.79ms ± 7%   1.77ms ± 6%      ~     (p=0.853 n=10+10)
SQL/Cockroach/Insert/count=1000-24                                                   10.6ms ± 6%   10.4ms ± 9%      ~     (p=0.393 n=10+10)
SQL/Cockroach/InsertDistinct/count=1-24                                              1.66ms ± 4%   1.63ms ± 4%      ~     (p=0.075 n=10+10)
SQL/Cockroach/InsertDistinct/count=10-24                                              326µs ± 7%    333µs ±10%      ~     (p=0.342 n=10+10)
SQL/Cockroach/InsertDistinct/count=100-24                                             346µs ± 3%    345µs ± 6%      ~     (p=0.912 n=10+10)
SQL/Cockroach/InsertDistinct/count=1000-24                                            817µs ±14%    881µs ±14%      ~     (p=0.315 n=10+8)
SQL/Cockroach/InsertFK/count=1/nFks=1-24                                              746µs ± 3%    754µs ± 8%      ~     (p=0.684 n=10+10)
SQL/Cockroach/InsertFK/count=1/nFks=5-24                                             1.09ms ± 4%   1.07ms ± 3%      ~     (p=0.075 n=10+10)
SQL/Cockroach/InsertFK/count=1/nFks=10-24                                            1.37ms ± 3%   1.35ms ± 3%      ~     (p=0.123 n=10+10)
SQL/Cockroach/InsertFK/count=10/nFks=1-24                                            1.60ms ± 3%   1.60ms ± 4%      ~     (p=0.842 n=10+9)
SQL/Cockroach/InsertFK/count=10/nFks=5-24                                            4.40ms ± 3%   4.30ms ± 6%      ~     (p=0.105 n=10+10)
SQL/Cockroach/InsertFK/count=10/nFks=10-24                                           6.65ms ±11%   6.21ms ± 9%    -6.64%  (p=0.012 n=8+10)
SQL/Cockroach/InsertFK/count=100/nFks=1-24                                           8.75ms ± 1%   8.72ms ± 2%      ~     (p=1.000 n=8+10)
SQL/Cockroach/InsertFK/count=100/nFks=5-24                                           35.5ms ± 4%   35.1ms ± 5%      ~     (p=0.237 n=10+8)
SQL/Cockroach/InsertFK/count=100/nFks=10-24                                          53.4ms ± 8%   52.1ms ± 3%      ~     (p=0.211 n=10+9)
SQL/Cockroach/InsertFK/count=1000/nFks=1-24                                          80.8ms ± 8%   78.6ms ± 6%      ~     (p=0.190 n=10+10)
SQL/Cockroach/InsertFK/count=1000/nFks=5-24                                           345ms ± 4%    350ms ± 5%      ~     (p=0.497 n=10+9)
SQL/Cockroach/InsertFK/count=1000/nFks=10-24                                          539ms ± 4%    533ms ± 6%      ~     (p=0.353 n=10+10)
SQL/Cockroach/InsertSecondaryIndex/count=1-24                                        1.27ms ± 1%   1.24ms ± 3%    -2.29%  (p=0.004 n=9+9)
SQL/Cockroach/InsertSecondaryIndex/count=10-24                                       2.08ms ± 7%   2.11ms ± 4%      ~     (p=0.481 n=10+10)
SQL/Cockroach/InsertSecondaryIndex/count=100-24                                      8.05ms ± 4%   8.52ms ± 7%    +5.90%  (p=0.009 n=10+10)
SQL/Cockroach/InsertSecondaryIndex/count=1000-24                                     69.7ms ± 9%   74.1ms ±18%      ~     (p=0.165 n=10+10)
SQL/Cockroach/InterleavedSelect/count=1-24                                            486µs ± 1%    490µs ± 2%    +0.82%  (p=0.038 n=8+8)
SQL/Cockroach/InterleavedSelect/count=10-24                                           506µs ± 1%    500µs ± 1%    -1.03%  (p=0.009 n=10+10)
SQL/Cockroach/InterleavedSelect/count=100-24                                          657µs ± 2%    654µs ± 1%      ~     (p=0.182 n=9+10)
SQL/Cockroach/InterleavedSelect/count=1000-24                                        1.96ms ± 2%   1.96ms ± 2%      ~     (p=0.739 n=10+10)
SQL/Cockroach/InterleavedFK/count=1-24                                                803µs ± 3%    797µs ± 1%      ~     (p=0.123 n=10+10)
SQL/Cockroach/InterleavedFK/count=10-24                                              1.65ms ± 3%   1.62ms ± 3%      ~     (p=0.143 n=10+10)
SQL/Cockroach/InterleavedFK/count=100-24                                             9.10ms ± 2%   8.71ms ± 7%      ~     (p=0.101 n=8+10)
SQL/Cockroach/InterleavedFK/count=1000-24                                            80.1ms ± 6%   77.6ms ± 2%      ~     (p=0.065 n=10+9)
SQL/Cockroach/TrackChoices/count=1-24                                                1.00ms ±13%   0.99ms ±18%      ~     (p=0.853 n=10+10)
SQL/Cockroach/TrackChoices/count=10-24                                                149µs ±12%    150µs ±12%      ~     (p=0.905 n=10+9)
SQL/Cockroach/TrackChoices/count=100-24                                              48.6µs ± 6%   51.0µs ± 7%    +4.76%  (p=0.023 n=10+10)
SQL/Cockroach/TrackChoices/count=1000-24                                             38.4µs ± 4%   40.8µs ±11%    +6.17%  (p=0.002 n=10+10)
SQL/Cockroach/Update/count=1-24                                                       903µs ± 3%    917µs ± 3%    +1.62%  (p=0.024 n=9+9)
SQL/Cockroach/Update/count=10-24                                                     1.23ms ± 3%   1.23ms ± 4%      ~     (p=0.912 n=10+10)
SQL/Cockroach/Update/count=100-24                                                    3.53ms ± 1%   3.64ms ± 3%    +3.24%  (p=0.002 n=8+9)
SQL/Cockroach/Update/count=1000-24                                                   23.8ms ± 5%   24.7ms ± 6%    +4.14%  (p=0.008 n=9+10)
SQL/Cockroach/Upsert/count=1-24                                                      1.75ms ±11%   1.86ms ±13%      ~     (p=0.052 n=10+10)
SQL/Cockroach/Upsert/count=10-24                                                     1.86ms ± 1%   2.08ms ±36%      ~     (p=0.089 n=10+10)
SQL/Cockroach/Upsert/count=100-24                                                    3.59ms ± 3%   3.62ms ± 2%      ~     (p=0.105 n=10+10)
SQL/Cockroach/Upsert/count=1000-24                                                   25.6ms ± 5%   27.0ms ± 5%    +5.36%  (p=0.005 n=10+10)
SQL/MultinodeCockroach/Delete/count=1-24                                              920µs ±32%   1152µs ± 3%   +25.31%  (p=0.003 n=10+8)
SQL/MultinodeCockroach/Delete/count=10-24                                             937µs ± 2%   1233µs ± 3%   +31.59%  (p=0.000 n=9+9)
SQL/MultinodeCockroach/Delete/count=100-24                                           1.62ms ± 3%   1.97ms ±15%   +22.09%  (p=0.000 n=8+8)
SQL/MultinodeCockroach/Delete/count=1000-24                                          21.6ms ±11%   23.1ms ±12%      ~     (p=0.075 n=10+10)
SQL/MultinodeCockroach/Insert/count=1-24                                              872µs ± 4%   1159µs ± 2%   +32.94%  (p=0.000 n=10+10)
SQL/MultinodeCockroach/Insert/count=10-24                                            1.01ms ± 3%   1.32ms ± 4%   +31.05%  (p=0.000 n=9+8)
SQL/MultinodeCockroach/Insert/count=100-24                                           2.02ms ± 3%   2.37ms ± 4%   +17.60%  (p=0.000 n=9+10)
SQL/MultinodeCockroach/Insert/count=1000-24                                          10.7ms ± 7%   12.2ms ± 7%   +13.84%  (p=0.000 n=10+9)
SQL/MultinodeCockroach/InsertDistinct/count=1-24                                     1.97ms ± 7%   2.27ms ± 1%   +14.78%  (p=0.000 n=10+9)
SQL/MultinodeCockroach/InsertDistinct/count=10-24                                     361µs ± 4%    422µs ± 5%   +16.83%  (p=0.000 n=9+10)
SQL/MultinodeCockroach/InsertDistinct/count=100-24                                    366µs ± 4%    451µs ± 8%   +23.08%  (p=0.000 n=10+10)
SQL/MultinodeCockroach/InsertDistinct/count=1000-24                                   807µs ± 7%   1016µs ±51%      ~     (p=0.497 n=9+10)
SQL/MultinodeCockroach/InsertFK/count=1/nFks=1-24                                    1.01ms ± 1%   1.59ms ± 2%   +56.74%  (p=0.000 n=9+8)
SQL/MultinodeCockroach/InsertFK/count=1/nFks=5-24                                    1.35ms ± 4%   2.00ms ± 3%   +48.88%  (p=0.000 n=9+9)
SQL/MultinodeCockroach/InsertFK/count=1/nFks=10-24                                   1.64ms ± 2%   2.40ms ± 2%   +45.89%  (p=0.000 n=10+9)
SQL/MultinodeCockroach/InsertFK/count=10/nFks=1-24                                   1.93ms ± 4%   4.99ms ± 2%  +158.86%  (p=0.000 n=9+10)
SQL/MultinodeCockroach/InsertFK/count=10/nFks=5-24                                   4.49ms ± 2%   8.45ms ±13%   +88.26%  (p=0.000 n=8+10)
SQL/MultinodeCockroach/InsertFK/count=10/nFks=10-24                                  6.42ms ± 3%  10.88ms ± 2%   +69.50%  (p=0.000 n=10+8)
SQL/MultinodeCockroach/InsertFK/count=100/nFks=1-24                                  9.43ms ± 1%  37.78ms ± 4%  +300.59%  (p=0.000 n=10+10)
SQL/MultinodeCockroach/InsertFK/count=100/nFks=5-24                                  34.9ms ± 5%   68.7ms ± 3%   +96.76%  (p=0.000 n=10+9)
SQL/MultinodeCockroach/InsertFK/count=100/nFks=10-24                                 53.6ms ± 4%   90.4ms ± 1%   +68.65%  (p=0.000 n=10+8)
SQL/MultinodeCockroach/InsertFK/count=1000/nFks=1-24                                 84.4ms ± 3%  373.2ms ± 3%  +342.25%  (p=0.000 n=10+10)
SQL/MultinodeCockroach/InsertFK/count=1000/nFks=5-24                                  359ms ± 9%    803ms ± 2%  +123.65%  (p=0.000 n=9+7)
SQL/MultinodeCockroach/InsertFK/count=1000/nFks=10-24                                 551ms ± 4%    967ms ±18%   +75.44%  (p=0.000 n=10+10)
SQL/MultinodeCockroach/InsertSecondaryIndex/count=1-24                               1.64ms ± 5%   2.02ms ± 3%   +23.08%  (p=0.000 n=10+9)
SQL/MultinodeCockroach/InsertSecondaryIndex/count=10-24                              2.58ms ± 8%   2.83ms ± 6%    +9.91%  (p=0.000 n=9+10)
SQL/MultinodeCockroach/InsertSecondaryIndex/count=100-24                             9.90ms ± 9%  10.67ms ±10%    +7.81%  (p=0.007 n=10+10)
SQL/MultinodeCockroach/InsertSecondaryIndex/count=1000-24                            80.7ms ± 8%   88.2ms ±15%    +9.37%  (p=0.010 n=10+9)
SQL/MultinodeCockroach/InterleavedSelect/count=1-24                                   521µs ± 1%   1041µs ± 2%   +99.79%  (p=0.000 n=10+9)
SQL/MultinodeCockroach/InterleavedSelect/count=10-24                                  535µs ± 1%   1088µs ± 3%  +103.48%  (p=0.000 n=7+10)
SQL/MultinodeCockroach/InterleavedSelect/count=100-24                                 704µs ± 2%   1299µs ± 9%   +84.64%  (p=0.000 n=10+10)
SQL/MultinodeCockroach/InterleavedSelect/count=1000-24                               2.10ms ± 3%   2.81ms ± 5%   +33.77%  (p=0.000 n=10+10)
SQL/MultinodeCockroach/InterleavedFK/count=1-24                                      1.32ms ± 3%   2.27ms ± 8%   +72.51%  (p=0.000 n=9+10)
SQL/MultinodeCockroach/InterleavedFK/count=10-24                                     2.52ms ± 5%   6.88ms ± 4%  +172.53%  (p=0.000 n=10+10)
SQL/MultinodeCockroach/InterleavedFK/count=100-24                                    13.8ms ± 6%   53.4ms ± 2%  +285.97%  (p=0.000 n=10+10)
SQL/MultinodeCockroach/InterleavedFK/count=1000-24                                    126ms ±17%    520ms ± 5%  +312.42%  (p=0.000 n=10+8)
SQL/MultinodeCockroach/TrackChoices/count=1-24                                       1.53ms ±25%   1.62ms ±11%      ~     (p=0.436 n=10+10)
SQL/MultinodeCockroach/TrackChoices/count=10-24                                       202µs ±45%    207µs ± 8%      ~     (p=0.796 n=10+10)
SQL/MultinodeCockroach/TrackChoices/count=100-24                                     53.4µs ±13%   57.6µs ± 6%    +7.74%  (p=0.043 n=10+10)
SQL/MultinodeCockroach/TrackChoices/count=1000-24                                    42.1µs ± 4%   41.4µs ± 4%      ~     (p=0.101 n=8+10)
SQL/MultinodeCockroach/Update/count=1-24                                             1.19ms ± 2%   1.80ms ±11%   +50.85%  (p=0.000 n=9+10)
SQL/MultinodeCockroach/Update/count=10-24                                            1.51ms ± 2%   2.18ms ± 2%   +44.59%  (p=0.000 n=8+9)
SQL/MultinodeCockroach/Update/count=100-24                                           3.81ms ± 4%   4.95ms ± 8%   +29.74%  (p=0.000 n=10+9)
SQL/MultinodeCockroach/Update/count=1000-24                                          24.0ms ± 4%   27.6ms ± 2%   +15.18%  (p=0.000 n=9+8)
SQL/MultinodeCockroach/Upsert/count=1-24                                             2.52ms ± 2%   3.02ms ± 2%   +19.80%  (p=0.000 n=9+10)
SQL/MultinodeCockroach/Upsert/count=10-24                                            3.60ms ± 3%   4.07ms ± 8%   +13.19%  (p=0.000 n=8+10)
SQL/MultinodeCockroach/Upsert/count=100-24                                           13.8ms ± 1%   11.9ms ± 1%   -13.93%  (p=0.000 n=10+9)
SQL/MultinodeCockroach/Upsert/count=1000-24                                          66.0ms ±92%  106.6ms ± 2%      ~     (p=0.497 n=10+9)
Scan/Cockroach/count=1/limit=0-24                                                     372µs ± 2%    377µs ± 1%    +1.37%  (p=0.019 n=10+10)
Scan/Cockroach/count=1/limit=1-24                                                     379µs ± 1%    388µs ± 2%    +2.56%  (p=0.000 n=9+9)
Scan/Cockroach/count=1/limit=10-24                                                    380µs ± 1%    388µs ± 2%    +2.08%  (p=0.001 n=9+10)
Scan/Cockroach/count=1/limit=100-24                                                   382µs ± 2%    394µs ± 1%    +2.99%  (p=0.000 n=10+9)
Scan/Cockroach/count=10/limit=0-24                                                    390µs ± 1%    400µs ± 3%    +2.41%  (p=0.000 n=10+10)
Scan/Cockroach/count=10/limit=1-24                                                    386µs ± 2%    393µs ± 2%    +1.83%  (p=0.002 n=10+8)
Scan/Cockroach/count=10/limit=10-24                                                   399µs ± 1%    406µs ± 2%    +1.92%  (p=0.000 n=10+9)
Scan/Cockroach/count=10/limit=100-24                                                  398µs ± 2%    410µs ± 1%    +2.83%  (p=0.000 n=8+10)
Scan/Cockroach/count=100/limit=0-24                                                   512µs ± 1%    526µs ± 5%    +2.72%  (p=0.007 n=10+10)
Scan/Cockroach/count=100/limit=1-24                                                   385µs ± 2%    394µs ± 2%    +2.21%  (p=0.000 n=10+10)
Scan/Cockroach/count=100/limit=10-24                                                  413µs ±10%    409µs ± 1%      ~     (p=0.353 n=10+10)
Scan/Cockroach/count=100/limit=100-24                                                 518µs ± 1%    522µs ± 1%    +0.92%  (p=0.003 n=8+10)
Scan/Cockroach/count=1000/limit=0-24                                                 1.57ms ± 1%   1.58ms ± 1%    +0.54%  (p=0.003 n=10+10)
Scan/Cockroach/count=1000/limit=1-24                                                  383µs ± 1%    394µs ± 1%    +2.93%  (p=0.000 n=9+10)
Scan/Cockroach/count=1000/limit=10-24                                                 398µs ± 2%    408µs ± 2%    +2.67%  (p=0.000 n=10+8)
Scan/Cockroach/count=1000/limit=100-24                                                520µs ± 1%    525µs ± 1%    +1.07%  (p=0.004 n=10+8)
Scan/Cockroach/count=10000/limit=0-24                                                10.4ms ± 2%   10.3ms ± 2%      ~     (p=0.105 n=8+8)
Scan/Cockroach/count=10000/limit=1-24                                                 400µs ± 3%    408µs ± 2%    +1.95%  (p=0.008 n=9+10)
Scan/Cockroach/count=10000/limit=10-24                                                423µs ± 3%    425µs ± 3%      ~     (p=0.481 n=10+10)
Scan/Cockroach/count=10000/limit=100-24                                               566µs ± 5%    557µs ± 1%      ~     (p=0.546 n=9+9)
Scan/MultinodeCockroach/count=1/limit=0-24                                            386µs ± 1%    848µs ± 3%  +119.77%  (p=0.000 n=10+10)
Scan/MultinodeCockroach/count=1/limit=1-24                                            393µs ± 2%    858µs ± 3%  +118.11%  (p=0.000 n=10+10)
Scan/MultinodeCockroach/count=1/limit=10-24                                           398µs ± 1%    858µs ± 3%  +115.68%  (p=0.000 n=10+8)
Scan/MultinodeCockroach/count=1/limit=100-24                                          397µs ± 1%    860µs ± 2%  +116.91%  (p=0.000 n=10+9)
Scan/MultinodeCockroach/count=10/limit=0-24                                           400µs ± 1%    868µs ± 2%  +117.05%  (p=0.000 n=10+9)
Scan/MultinodeCockroach/count=10/limit=1-24                                           397µs ± 3%    865µs ± 3%  +117.95%  (p=0.000 n=10+9)
Scan/MultinodeCockroach/count=10/limit=10-24                                          408µs ± 1%    879µs ± 2%  +115.53%  (p=0.000 n=9+10)
Scan/MultinodeCockroach/count=10/limit=100-24                                         410µs ± 1%    879µs ± 2%  +114.30%  (p=0.000 n=9+10)
…